CHROMIUM: iwl7000: dbg: don't abort if sending DBGC_SUSPEND_RESUME fails
If the firmware is in a bad state or not initialized fully, sending
the DBGC_SUSPEND_RESUME command fails but we can still collect logs.

Instead of aborting the entire dump process, simply ignore the error.
By removing the last callpoint that was checking the return value, we
can also convert the function to return void.

Change-Id: I8dcd190431d110a6a0e88095ce93591ccfb3d78d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 5bcb625c51307e910803414300344d1793d0c2d2