CHROMIUM: iwl7000: mac80211: always set pwr_level_cisco to 0
We removed the code that handles CCX power levels, but we kept the
pwr_level_cisco variable and the checks.  We will never access
pwr_level_cisco, because we check for !has_cisco_pwr first and that is
always true in our code.  But some compilers don't understand this and
cause a warning.

Set the pwr_level_cisco variable to 0 (whatever value would be fine)
to avoid the warning.

Change-Id: I652bf844f4d0388a6a021ed1b3f0fdeb4e7e35f4
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: a63648fc53647d7355521a2474657f0086c6db49