CHROMIUM: iwl7000: pcie: gen2: use DMA pool for byte-count tables
Since the recent patch in this area, we no longer allocate 64k
for a single queue, but only 1k, which still means a full page.
Use a DMA pool to reduce this further, since we will have a lot
of queues in a typical system that can share pages.

Change-Id: Ie9a417132812d110ec1cc87852f101477c01cfcb
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: e1b145053efcac481cc2cda73ce2798650181b95