BACKPORT: ANDROID: FROMLIST: binder: remove binder_debug_no_lock mechanism
[Android 4.4. upstream a19f3efc4f54593d9e17f3fe7afc1ea78e0ae5f0]

(from https://patchwork.kernel.org/patch/9817811/)

With the global lock, there was a mechanism to acceess
binder driver debugging information with the global
lock disabled to debug deadlocks or other issues.
This mechanism is rarely (if ever) used anymore
and wasn't needed during the development of
fine-grained locking in the binder driver.
Removing it.

BUG=b:142440806
Change-Id: I9504d35b9c8a9cb0ccbbeaf257cd312edb76be3f

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>