BACKPORT: ANDROID: FROMLIST: binder: refactor queue management in binder_thread_read
[Android 4.4. upstream 9b9340c58afa2cf2a76ae06f3019ad631e6f7799]

(from https://patchwork.kernel.org/patch/9817757/)

In binder_thread_read, the BINDER_WORK_NODE command is used
to communicate the references on the node to userspace. It
can take a couple of iterations in the loop to construct
the list of commands for user space. When locking is added,
the lock would need to be release on each iteration which
means the state could change. The work item is not dequeued
during this process which prevents a simpler queue management
that can just dequeue up front and handle the work item.

Fixed by changing the BINDER_WORK_NODE algorithm in
binder_thread_read to determine which commands to send
to userspace atomically in 1 pass so it stays consistent
with the kernel view.

The work item is now dequeued immediately since only
1 pass is needed.

BUG=b:142440806
Change-Id: Ide377b3bde271eb4f8f6932760de516d3d1bb5fc

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>