BACKPORT: ANDROID: FROMLIST: binder: avoid race conditions when enqueuing txn
[Android 4.4 upstream 6ea6027115718c27e01e4459901c7938492cfa8f]

(from https://patchwork.kernel.org/patch/9817813/)

Currently, the transaction complete work item is queued
after the transaction. This means that it is possible
for the transaction to be handled and a reply to be
enqueued in the current thread before the transaction
complete is enqueued, which violates the protocol
with userspace who may not expect the transaction
complete. Fixed by always enqueing the transaction
complete first.

Also, once the transaction is enqueued, it is unsafe
to access since it might be freed. Currently,
t->flags is accessed to determine whether a sync
wake is needed. Changed to access tr->flags
instead.

BUG=b:142440806
Change-Id: Ic71df497e961a66f045e44572d00350e268a35eb

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>