BACKPORT: ANDROID: FROMLIST: binder: use atomic for transaction_log index
[Android 4.4 upstream 0f32aeb35f8b8c3d07f1c03062cc195590145d35]

(from https://patchwork.kernel.org/patch/9817807/)

The log->next index for the transaction log was
not protected when incremented. This led to a
case where log->next++ resulted in an index
larger than ARRAY_SIZE(log->entry) and eventually
a bad access to memory.

Fixed by making the log index an atomic64 and
converting to an array by using "% ARRAY_SIZE(log->entry)"

Also added "complete" field to the log entry which is
written last to tell the print code whether the
entry is complete

BUG=b:142440806
Change-Id: I05df90e24a48795557f8eb8c6cb5d757d9ef5868

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>