BACKPORT: ANDROID: FROMLIST: binder: refactor binder_pop_transaction
[Android 4.4 upstream 162735381739bb4f3e5ad6f35a6319d99ee7e697]

(from https://lkml.org/lkml/2017/6/29/754)

binder_pop_transaction needs to be split into 2 pieces to
to allow the proc lock to be held on entry to dequeue the
transaction stack, but no lock when kfree'ing the transaction.

Split into binder_pop_transaction_locked and binder_free_transaction
(the actual locks are still to be added).

BUG=b:142440806
Change-Id: Ib8acec59be85e043233b60831bb5c1e2c13d8fd3

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>