BACKPORT: ANDROID: FROMLIST: binder: guarantee txn complete / errors delivered in-order
[Android 4.4. upstream 3a822b33c84bf30d2081ede2c42d53ae7ab89d85]

(from https://patchwork.kernel.org/patch/9817805/)

Since errors are tracked in the return_error/return_error2
fields of the binder_thread object and BR_TRANSACTION_COMPLETEs
can be tracked either in those fields or via the thread todo
work list, it is possible for errors to be reported ahead
of the associated txn complete.

Use the thread todo work list for errors to guarantee
order. Also changed binder_send_failed_reply to pop
the transaction even if it failed to send a reply.

BUG=b:142440806
Change-Id: I12be420cad344ce395b4d7725568ca750d0c59b7

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>