BACKPORT: ANDROID: FROMLIST: binder: make sure accesses to proc/thread are safe
[Android 4.4 upstream e482ec39d635cd1bcf764f2aefbc4ab3149d2299]

(from https://patchwork.kernel.org/patch/9817787/)

binder_thread and binder_proc may be accessed by other
threads when processing transaction. Therefore they
must be prevented from being freed while a transaction
is in progress that references them.

This is done by introducing a temporary reference
counter for threads and procs that indicates that the
object is in use and must not be freed. binder_thread_dec_tmpref()
and binder_proc_dec_tmpref() are used to decrement
the temporary reference.

It is safe to free a binder_thread if there
is no reference and it has been released
(indicated by thread->is_dead).

It is safe to free a binder_proc if it has no
remaining threads and no reference.

A spinlock is added to the binder_transaction
to safely access and set references for t->from
and for debug code to safely access t->to_thread
and t->to_proc.

BUG=b:142440806
Change-Id: I75790d3318f18a70db9d57a807bb675bf528aad9

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>