BACKPORT: ANDROID: FROMLIST: binder: use inner lock to sync work dq and node counts
[Android 4.4 upstream f73f378b52e81acdb9654a5bf81661d9ae461e62]

(from https://patchwork.kernel.org/patch/9817789/)

For correct behavior we need to hold the inner lock when
dequeuing and processing node work in binder_thread_read.
We now hold the inner lock when we enter the switch statement
and release it after processing anything that might be
affected by other threads.

We also need to hold the inner lock to protect the node
weak/strong ref tracking fields as long as node->proc
is non-NULL (if it is NULL then we are guaranteed that
we don't have any node work queued).

This means that other functions that manipulate these fields
must hold the inner lock. Refactored these functions to use
the inner lock.

BUG=b:142440806
Change-Id: I049aa597ebb51d96f6b0c759274b474c3c4382df

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>