BACKPORT: ANDROID: FROMLIST: binder: protect transaction_stack with inner lock.
[Android 4.4 upstream 89b657e0d7de20459820d7edf341bc888b69aa83]

(from https://patchwork.kernel.org/patch/9817779/)

This makes future changes to priority inheritance
easier, since we want to be able to look at a thread's
transaction stack when selecting a thread to inherit
priority for.

It also allows us to take just a single lock in a
few paths, where we used to take two in succession.

BUG=b:142440806
Change-Id: I0777d11850a1b2aeee4ae19e22294159693fc2b5

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>