BACKPORT: ANDROID: FROMLIST: binder: protect against stale pointers in print_binder_transaction
[Android 4.4 upstream da957e45dd21fee345544dadf0fc426abe503d28]

(from https://patchwork.kernel.org/patch/9817761/)

When printing transactions there were several race conditions
that could cause a stale pointer to be deferenced. Fixed by
reading the pointer once and using it if valid (which is
safe). The transaction buffer also needed protection via proc
lock, so it is only printed if we are holding the correct lock.

BUG=b:142440806
Change-Id: I249ed02947956261fd22e1283af44c32f79e1904

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>