BACKPORT: ANDROID: ANDROID: binder: add min sched_policy to node.
[Android 4.4 upstream adb685439e768b120a5bdae15f333967d1e239ed]

This change adds flags to flat_binder_object.flags
to allow indicating a minimum scheduling policy for
the node. It also clarifies the valid value range
for the priority bits in the flags.

Internally, we use the priority map that the kernel
uses, e.g. [0..99] for real-time policies and [100..139]
for the SCHED_NORMAL/SCHED_BATCH policies.

BUG=b:142440806
Change-Id: I0143cb11da98766aaaf498f4dae378bbdb799454

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>