BACKPORT: ANDROID: FROMLIST: android: binder: Move buffer out of area shared with user space
[Android 4.4 upstream 7a6d4b157e1c046f4c4de9f2f1f1f7abfa6abb21]

(from https://patchwork.kernel.org/patch/9928607/)

Binder driver allocates buffer meta data in a region that is mapped
in user space. These meta data contain pointers in the kernel.

This patch allocates buffer meta data on the kernel heap that is
not mapped in user space, and uses a pointer to refer to the data mapped.

Also move alloc->buffers initialization from mmap to init since it's
now used even when mmap failed or was not called.

BUG=b:142440806
Change-Id: I7441b77227e64246bcfb79402d42a8ed7d466eb5

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>