BACKPORT: ANDROID: FROMLIST: binder: fix use-after-free in binder_transaction()
[Android 4.4 upstream 642da1dadeb8bdbeab8acd138afbdc041a723ae9]

(from https://patchwork.kernel.org/patch/9978801/)

User-space normally keeps the node alive when creating a transaction
since it has a reference to the target. The local strong ref keeps it
alive if the sending process dies before the target process processes
the transaction. If the source process is malicious or has a reference
counting bug, this can fail.

In this case, when we attempt to decrement the node in the failure
path, the node has already been freed.

This is fixed by taking a tmpref on the node while constructing
the transaction. To avoid re-acquiring the node lock and inner
proc lock to increment the proc's tmpref, a helper is used that
does the ref increments on both the node and proc.

BUG=b:142440806
Change-Id: I090d969b172230b106002581693e00bcd8ab8da2

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>