BACKPORT: ANDROID: FROMLIST: binder: fix proc->files use-after-free
[Android 4.4 upstream 19ef30ef23a4de3f0bd8979a5bb5216dd381bc75]

(from https://patchwork.kernel.org/patch/10058587/)

proc->files cleanup is initiated by binder_vma_close. Therefore
a reference on the binder_proc is not enough to prevent the
files_struct from being released while the binder_proc still has
a reference. This can lead to an attempt to dereference the
stale pointer obtained from proc->files prior to proc->files
cleanup. This has been seen once in task_get_unused_fd_flags()
when __alloc_fd() is called with a stale "files".

The fix is to always use get_files_struct() to obtain struct_files
so that the refcount on the files_struct is used to prevent
a premature free. proc->files is removed since we get it every
time.

BUG=b:142440806
Change-Id: I8ac917f5ac6463f24dfe046f44fc61a65f107d3c

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>