BACKPORT: ANDROID: UPSTREAM: ANDROID: binder: remove waitqueue when thread exits.
[Android 4.4 upstream 810bdaf1dd747016fea5d5256530cf9ade05a4c9]

binder_poll() passes the thread->wait waitqueue that
can be slept on for work. When a thread that uses
epoll explicitly exits using BINDER_THREAD_EXIT,
the waitqueue is freed, but it is never removed
from the corresponding epoll data structure. When
the process subsequently exits, the epoll cleanup
code tries to access the waitlist, which results in
a use-after-free.

Prevent this by using POLLFREE when the thread exits.

(cherry picked from commit f5cb779ba16334b45ba8946d6bfa6d9834d1527f)

BUG=b:142440806
Change-Id: I12c2acdde0460a7c71f2634e84e9137555031371

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>