BACKPORT: ANDROID: UPSTREAM: ANDROID: binder: re-order some conditions
[Android 4.4 upstream 02d82286734d9195913adf42300523c41f1008ab]

It doesn't make any difference to runtime but I've switched these two
checks to make my static checker happy.

The problem is that "buffer->data_size" is user controlled and if it's
less than "sizeo(*hdr)" then that means "offset" can be more than
"buffer->data_size".  It's just cleaner to check it in the other order.

BUG=b:142440806
Change-Id: Id3a374aaf7f3d23d176981ab84c3626fbeb8788c

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>