BACKPORT: ANDROID: UPSTREAM: binder: fix proc->files use-after-free
[Android 4.4 upstream c88a3ec1ee64576a706432af3e8f1aee8e6e4d2e]

proc->files cleanup is initiated by binder_vma_close. Therefore
a reference on the binder_proc is not enough to prevent the
files_struct from being released while the binder_proc still has
a reference. This can lead to an attempt to dereference the
stale pointer obtained from proc->files prior to proc->files
cleanup. This has been seen once in task_get_unused_fd_flags()
when __alloc_fd() is called with a stale "files".

The fix is to protect proc->files with a mutex to prevent cleanup
while in use.

BUG=b:142440806
Change-Id: I00d87b58e09836ba4208572a490652979f2ee8da

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>