BACKPORT: ANDROID: UPSTREAM: binder: fix race that allows malicious free of live buffer
[Android 4.4 upstream 8250e54af25dadfe8de0a86ca0d3a60d9da7b8dc]

commit 7bada55ab50697861eee6bb7d60b41e68a961a9c upstream

Malicious code can attempt to free buffers using the BC_FREE_BUFFER
ioctl to binder. There are protections against a user freeing a buffer
while in use by the kernel, however there was a window where
BC_FREE_BUFFER could be used to free a recently allocated buffer that
was not completely initialized. This resulted in a use-after-free
detected by KASAN with a malicious test program.

This window is closed by setting the buffer's allow_user_free attribute
to 0 when the buffer is allocated or when the user has previously freed
it instead of waiting for the caller to set it. The problem was that
when the struct buffer was recycled, allow_user_free was stale and set
to 1 allowing a free to go through.

BUG=b:142440806
Change-Id: I51269d78602a5bc26695ce80c9d66166ef529a57

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>