BACKPORT: ANDROID: FROMGIT: binder: create node flag to request sender's security context
[Android 4.4 upstream 181243d59e981b2bb3e5b88d9e261f316e4583a1]

To allow servers to verify client identity, allow a node
flag to be set that causes the sender's security context
to be delivered with the transaction. The BR_TRANSACTION
command is extended in BR_TRANSACTION_SEC_CTX to
contain a pointer to the security context string.

Signed-off-by: Todd Kjos <tkjos@google.com>
Reviewed-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit ec74136ded792deed80780a2f8baf3521eeb72f9
 https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
 master)

BUG=b:142440806
Change-Id: If1c6e46bf70456232c29a31307630fbd8013fd89

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>