BACKPORT: ANDROID: BACKPORT: binder: fix race between munmap() and direct reclaim
[Android 4.4 upstream 8236fb88a8c1ff2dfd9ac538ac54f8fbba2ade63]

commit 5cec2d2e5839f9c0fec319c523a911e0a7fd299f upstream.

An munmap() on a binder device causes binder_vma_close() to be called
which clears the alloc->vma pointer.

If direct reclaim causes binder_alloc_free_page() to be called, there
is a race where alloc->vma is read into a local vma pointer and then
used later after the mm->mmap_sem is acquired. This can result in
calling zap_page_range() with an invalid vma which manifests as a
use-after-free in zap_page_range().

The fix is to check alloc->vma after acquiring the mmap_sem (which we
were acquiring anyway) and skip zap_page_range() if it has changed
to NULL.

BUG=b:142440806
Change-Id: If4b6781c364d647fcc9f8c2579ecd081d0e6af70

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>