BACKPORT: UPSTREAM: kbuild: Disable -Wpointer-to-enum-cast
Clang's -Wpointer-to-int-cast deviates from GCC in that it warns when
casting to enums. The kernel does this in certain places, such as device
tree matches to set the version of the device being used, which allows
the kernel to avoid using a gigantic union.

https://elixir.bootlin.com/linux/v5.5.8/source/drivers/ata/ahci_brcm.c#L428
https://elixir.bootlin.com/linux/v5.5.8/source/drivers/ata/ahci_brcm.c#L402
https://elixir.bootlin.com/linux/v5.5.8/source/include/linux/mod_devicetable.h#L264

To avoid a ton of false positive warnings, disable this particular part
of the warning, which has been split off into a separate diagnostic so
that the entire warning does not need to be turned off for clang. It
will be visible under W=1 in case people want to go about fixing these
easily and enabling the warning treewide.

Cc: stable@vger.kernel.org
Link: https://github.com/ClangBuiltLinux/linux/issues/887
Link: https://github.com/llvm/llvm-project/commit/2a41b31fcdfcb67ab7038fc2ffb606fd50b83a84
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
(cherry picked from commit 82f2bc2fcc0160d6f82dd1ac64518ae0a4dd183f)

BUG=chromium:1058965
TEST=kernel compiles with ToT clang

Signed-off-by: Manoj Gupta <manojgupta@google.com>
Change-Id: I54af136f946e72d61ab0874c6fa0498b6bb585e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2103960
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>