CHROMIUM: iwl7000: mvm: fix aux station leak
When mvm is initialized we alloc aux station with aux queue.
We later free the station memory when driver is stopped, but we
never free the queue's memory, which casues a leak.

Add a proper de-initialization of the station.

Change-Id: Id7516fbb3482131d0c9dfb51ff20b226617ddb49
Signed-off-by: Sharon <sara.sharon@intel.com>
iwl7000-tree: cc01cff0d75553ce7905e019d72f5cdf960e4f66