BACKPORT: media: Kconfig files: use select for V4L2 subdevs and MC
There are lots of drivers that only work when the media controller
and/or the V4L2 subdev APIs are present.

Right now, someone need to first enable those APIs before
using those drivers.

Well, ideally, drivers, should, instead *optionally*
depend on it, in order for PC camera drivers to be able to use
them, but nowadays most drivers are UVC cameras, with don't
require a sensor driver.

So, be it.

Let's instead make them select the MEDIA_CONTROLLER and the
SUBDEV API, in order to make easier for people to be able
of enabling them.

[ribalda: Keep only rkisp1 changes]

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 32a363d0b0b142f35512848dc646ee53e0926723)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I3c12e2267f8a059dd72f1db515d9647c62d3cadb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2517846
Reviewed-by: Tomasz Figa <tfiga@chromium.org>