CHROMIUM: iwl7000: msix: limit max RX queues for 9000 family
There is an issue in the HW DMA engine in the 9000 family of devices
when more than 6 RX queues are used.  The issue is that the FW may
hang when IWL_MVM_RXQ_NSSN_SYNC notifications are sent.

Fix this by limiting the number of RX queues to 6 in the 9000 family
of devices.

Change-Id: I8dfe8a7d3a7ac9f0bc9d93e4a03f8165d8c999d2
Signed-off-by: Mordechay Goodstein <mordechay.goodstein@intel.com>
iwl7000-tree: d6a036b834ccb9a671a3abd8e77bff512f74f9a0