CHROMIUM: iwl7000: chromiumOS: add set_wdev_cac_started
We assign a value to wdev->cac_started, so we can't make use of the
existing function, which only returns the value.  To solve this, add a
function to set it as well and use the adjustments spatch to convert
assignments to use it.

We don't really need this because the code that assigns values to it
is only used in >= 5.5.0, but we need to prevent the existing rule
from overwriting it.

Change-Id: Iff6d80524184482a7721300da4632d22cbfbc11f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: fc46bb4a21739ca7d26546a21a4efc7d1ffc60ef