CHROMIUM: fs/namespace.c: Copy nosymfollow to flags
The nosymfollow mount option used to be implemented in the chromiumos
LSM.  Since it wasn't a recognized flag, it was passed in the data
parameter to mount() and parsed out by the LSM.  Now that
crrev.com/c/2121041 adds a MNT_NOSYMFOLLOW option, symlink blocking is
moving over to this flag instead of the LSM.  For backwards
compatibility, we need to set the flag if nosymfollow is in the string
options until all callers are updated.

In addition, we want to start logging a warning when callers pass the
old flag so that they can be found and updated.  Some callers will pass
the flag both ways for cross-kernel compatibility, so only check/log the
warning if the new flag is not present.

This CL is essentially crrev.com/c/2125239 and crrev.com/c/2125239
squashed together.

BUG=b:152074038
TEST=tast run security.StatefulPartitionHardening
Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>
Change-Id: I5462550f7be69b80ac51f691828d590b2435276b
Reviewed-by: Ross Zwisler <zwisler@google.com>
(cherry picked from commit 4531e0a35c45e55fc63472e7d13178e2fbbb5077)
(cherry picked from commit 7210c76836224b3a5e915f10f337b726a9e0cfae)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2298465