CHROMIUM: iwl7000: mvm: add support for range request command ver 11
Version 11 of the range request command adds support for setting
the PN for secure ranging. For now, this is not yet supported.

The same functions that are used for version 9 and 10 are also
used for version 11 as the common part is the same.

Change-Id: I046e0d9f6dfaafda9794e5eb2ee1f02fcad2851a
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: eee801d95e5898129a4ff7f4ad83e21340497dc8