CHROMIUM: iwl7000: acpi: remove dummy definition of iwl_sar_set_profile()
This function is only called from acpi.c, which is only included in
the makefile if CONFIG_ACPI is set.  So it doesn't make sense to
declare a dummy version of it when CONFIG_ACPI is not defined.

Change-Id: I3271f9d2f7e484779a6319a1514cd0b7221fe326
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 14dbb809bfa5c756eb0a0c37420f84a32d265c10