CHROMIUM: drm/mst: "Fix" the race between detect and connector destroy
There's a race between detect and connector destroy causing a
kernel
OOPS. The race involves the port->kref.

detect    |    destroy_connector_work
----------+--------------------------
get       |
          |  kref_init
put       |
          |  put

The put in detect() will cause the destroy_worker to be
rescheduled for
the same port resulting in a double-free and an OOPS.

To fix, we serialize detect and destroy_work such that this can't
happen.

CHROMIUM because upstream has different fixes.

BUG=b:144866969
TEST=Connect 2 displays to eve via MST hub on 83.13020.54.0,
     unplug and observe
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I91cc5e7e6b4686a8b4d690027af5f2b5949acb00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2222915
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>