CHROMIUM: x86/vdso: Remove retpoline flags
Older compilers will sometimes generate an indirect call in the VDSO
which leads to a linker failure because it tries to use the
RETPOLINE.  This issue has ocurred on distro kernels previously and
this fix was proposed by Borislav Petkov <bp@suse.de>

see https://lore.kernel.org/patchwork/patch/879138/

So on those older vdsos, the compiler can generate indirect calls to
the retpoline thunks, which manifests itself as:

/usr/lib64/gcc/x86_64-suse-linux/4.8/../../../../x86_64-suse-linux/bin/ld:\
arch/x86/vdso/vclock_gettime.o: relocation R_X86_64_PC32 against \
undefined symbol `__x86_indirect_thunk_rax' can not be used when \
making a shared object; recompile with -fPIC

However, since the vdso is not part of the kernel, it cannot reference
kernel symbols so we better not build with with retpoline CFLAGS
either.

Thus, export RETPOLINE_CFLAGS to the lower level Makefiles so that
they can be filtered out in vdso's Makefile.

BUG=b:156315202
TEST=build amd64-generic-kernel-4_4 with USE="-clang -ldd"

Suggested-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Borislav Petkov <bp@suse.de>
Cc: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: I907a30c33b52a1ea1bbca6fc543c1275557073b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2233347
Reviewed-by: Jesse Barnes <jsbarnes@google.com>
Tested-by: Jesse Barnes <jsbarnes@google.com>