CHROMIUM: iwl7000: mvm: read and parse SKU ID if available
In v5 of the ALIVE notification, we may receive the SKU ID of the
device in use.  Read and store it locally for future access.  This
will be needed to find the correct PNVM (platform NVM) to pass to the
firmware.

Change-Id: Ibb0d88586a34acc904bf319974d791faafb7a905
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 8d3b5087dac5c6b738b97b22db0549d2fe1fa71b