CHROMIUM: iwl7000: mvm: don't send a CSA command the firmware doesn't know
We introduced a new flow to remove an on-going CSA but we
assumed the firwmare understands the
CHANNEL_SWITCH_TIME_EVENT_CMD. This is not true for 7265 and down.

Don't send this command for those devices, but rather use the older
command.

Change-Id: Ic0c56411da1096e9a45ee5c3383edf777c5baca0
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 60c262c0b346b347b4b03f82f54d33ecde279efa