CHROMIUM: iwl7000: read and parse PNVM file
The driver looks for a PNVM file that contains FW configuration data
for each different HW combination.  The FW requests the data for a
certain SKU_ID and the driver tries to find it in the PNVM file.

Read the file, parse its contents and send it to the trans.

Change-Id: I1d93dd6e6651586878db57fac3e7c3f09d742c42
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e1c502d15862f1673f4ab80318109553719d6002