CHROMIUM: iwl7000: mvm: validate firmware sync response size
We send some data to the firmware and expect to get it back,
but we shouldn't really trust the firmware on this. Check the
size of all the data we send down to avoid using bad or just
uninitialized data when the firmware doesn't respond right.

Change-Id: I4fa68bb2b1c7dcc52ddd381c4042722d27c4a34d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 457576ca2e36c2c8c33825b0faa8342d194b2701