CHROMIUM: iwl7000: mvm: scan: fix scheduled scan restart handling
When restarting firmware with an ongoing scheduled scan, we
don't (and shouldn't) mark it as aborted as mac80211 will be
restarting it, and so no event should go out to userspace.
The appropriate comment regarding this wasn't moved to this
place, so add it.

However, we _do_ need to clean up our internal state, since
mac80211 will restart the scan, and we'll otherwise get to
the WARN_ON() a few lines below for no reason whatsoever.

Change-Id: Ie869b628ae56a5d776eba0e7b7f05f42fc566f2e
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 56500a6a44639004e0a6fe5c03d44b3f03c247ad