UPSTREAM: media: staging: rkisp1: rsz: enumerate the formats on the src pad according to the capture
Currently the resizer outputs the same media bus format
as the input. This is wrong since the resizer is also used
to downscale YUV formats. This patch changes the enumeration
of the supported formats. The supported formats on the sink pad
should be taken from the isp entity and the supported formats on
the source pad should be taken from the capture entity.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit ca5331cf9d78c968e5cf28e4131765f869840817)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I2e0c8ce2177022a4eec7dfcc7e3b2cc655e7fd70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2517827
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>